Remove testingUser module
This module was doing exactly no-one good. It was broken, in a different
language than the core library, hard to package, and didn't do half of the
things that it was supposed to. It may come back in its own repo but for now it
is best to simply remove it.