Rewrite HttpRequest to rely more on OkHttp's HttpUrl
A few reasons I think this is a better design:

 1. Previously, calls to url() overrode any other URL-related properties,
    even if those properties were specified after the entire URL. This can
    lead to confusion since this is not the general behavior of the builder
    pattern.
 2. We don't have to reinvent the wheel when it comes to building a URL. Let
    the fine devs at Square handle that logic.
 3. HttpRequest can now have a newBuilder() method without the messy logic of
    deciding whether to specify the entire URL of its individual parts

Methods that modify the URL are deprecated in favor of the new url(HttpUrl)
method, except for endpoint(). That is a special method which HttpUrl does
not support, and will be kept for the time being. The newly deprecated
methods simply mess with the private HttpUrl.Builder instance.

I've also added a headers(okhttp3.Headers) method to specify all headers at
once.